requirements.txt
attrs==20.2.0
coverage==5.3
iniconfig==1.1.1
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pandasql==0.7.3
pathlib2==2.3.5
pep8==1.7.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.1
